add arguments for parallel execution of valgrind check

+++ scripts/check_valgrind.sh
@@ -23,7 +23,9 @@
 #
 # This script runs Valgrind on an input binary paths with specified memory leak detection flags. 
 # The Valgrind output is parsed. If any errors are found, they are printed and the script exits with a status of 1.
+# If errors are found, the Valgrind output is kept in the file valgrind-output.log for further inspection.
 # As a second argument, you can provide a path to a suppression file that is used by Valgrind to suppress certain errors.
+# As a third argument, you can provide the number of processes to use with mpi (default is 1).
 #
 
 # Check that an argument is given and that the argument is a file.
@@ -43,19 +45,8 @@ else
   fi
 fi
 
-# Write valgrind output to variable OUTPUT_FILE.
-OUTPUT_FILE="valgrind-output.log"
-# Set valgrind flags.
-VALGRIND_FLAGS="--leak-check=full --track-origins=yes \
-    --trace-children=yes --show-leak-kinds=definite,indirect,possible \
-    --errors-for-leak-kinds=definite,indirect,possible"
-# There are some more flags that can be reasonable to use, e.g., for debugging reasons if you found an error.
-# We used minimal flags for performance reasons.
-# Further flags include (but of course are not limited to): --expensive-definedness-checks=yes --track-fds=yes
-# For more detailed outputs: -read-var-info=yes --read-inline-info=yes --gen-suppressions=all
-# Warning: --show-leak-kinds=all will find a lot of still reachable leaks. This is not necessarily a problem.
-
 # Check if a second argument is provided. If yes, add the flag to incorporate the Valgrind suppression file.
+VALGRIND_FLAGS=""
 if ! [ ${2-x} = x ]; then
   if [ -f "$2" ]; then
     VALGRIND_FLAGS="${VALGRIND_FLAGS} --suppressions=${2}"
@@ -65,8 +56,27 @@ if ! [ ${2-x} = x ]; then
   fi
 fi
 
+# Check if a number of processes is provided as third argument. If not, set to 1.
+if [ -n "$3" ] && [[ "$3" =~ ^[0-9]+$ ]]; then
+  num_procs="$3"
+else
+  num_procs=1
+fi
+
+# Write valgrind output to variable OUTPUT_FILE.
+OUTPUT_FILE="valgrind-output.log"
+# Set valgrind flags.
+VALGRIND_FLAGS="${VALGRIND_FLAGS} --leak-check=full --track-origins=yes \
+    --trace-children=yes --show-leak-kinds=definite,indirect,possible \
+    --errors-for-leak-kinds=definite,indirect,possible"
+# There are some more flags that can be reasonable to use, e.g., for debugging reasons if you found an error.
+# We used minimal flags for performance reasons.
+# Further flags include (but of course are not limited to): --expensive-definedness-checks=yes --track-fds=yes
+# For more detailed outputs: -read-var-info=yes --read-inline-info=yes --gen-suppressions=all
+# Warning: --show-leak-kinds=all will find a lot of still reachable leaks. This is not necessarily a problem.
+
 # Run valgrind on given file with flags and write output to OUTPUT_FILE.
-valgrind $VALGRIND_FLAGS "${FILE}" > /dev/null 2>"${OUTPUT_FILE}"
+mpirun -n $num_procs valgrind $VALGRIND_FLAGS "${FILE}" > /dev/null 2>"${OUTPUT_FILE}"
 
 # Parse valgrind output.
 declare -a VALGRIND_RULES=(
@@ -82,6 +92,7 @@ declare -a VALGRIND_RULES=(
         "^==.*== Argument .* of function .* has a fishy (possibly negative) value: .*$"
         "^==.*== .*alloc() with size 0$"
         "^==.*== Invalid alignment value: .* (should be power of 2)$"
+        "^==.*== Conditional jump or move depends on uninitialised value(s)"
     )
 report_id=1
 status=0
@@ -122,5 +133,7 @@ while IFS= read -r line; do
   fi
 done < "${OUTPUT_FILE}"
 
-rm -f "${OUTPUT_FILE}"
+if [ "$status" -eq 0 ]; then
+  rm -f "${OUTPUT_FILE}"
+fi
 exit "${status}"
